+++ semantic_copycat_oslili/__init__.py
@@ -13,7 +13,7 @@ if os.environ.get('OSLILI_DEBUG') != '1':
     except ImportError:
         pass
 
-__version__ = "1.3.5"
+__version__ = "1.3.6"
 
 from .core.generator import LicenseCopyrightDetector
 from .core.models import (